requirements.txt

python-telegram-bot==20.7
Flask==2.3.3
pandas==2.2.2
openpyxl==3.1.2
gunicorn==21.2.0
